fix(sensors): widen octal-decode arithmetic to avoid u8 overflow
In decode_octal_escapes, the arithmetic was operating entirely in u8,
which caused overflow when bytes[i+1] >= '4' (e.g., (4..=7) * 64).
This would panic in debug builds and wrap silently in release.

Widen the arithmetic to u32 using u32::from() for each byte
subtraction, compute the full value, then truncate back to u8 with
a cast_possible_truncation allow (safe because mountinfo only emits
\000–\377, guaranteed <= 255).

+++ crates/hytte-services/src/sensors.rs
@@ -799,10 +799,11 @@ fn decode_octal_escapes(s: &str) -> String {
             && is_octal(bytes[i + 2])
             && is_octal(bytes[i + 3])
         {
-            let v = (bytes[i + 1] - b'0') * 64
-                + (bytes[i + 2] - b'0') * 8
-                + (bytes[i + 3] - b'0');
-            out.push(v);
+            let v = u32::from(bytes[i + 1] - b'0') * 64
+                + u32::from(bytes[i + 2] - b'0') * 8
+                + u32::from(bytes[i + 3] - b'0');
+            #[allow(clippy::cast_possible_truncation)]
+            out.push(v as u8);  // safe: mountinfo only emits \000–\377
             i += 4;
         } else {
             out.push(b);
